refactor storage backends into better arrangement
